固有ページの分離
