+++ i18n/ja.json
@@ -6,6 +6,7 @@
     "password": "パスワード"
   },
   "error": {
+    "api": "処理実行中にエラーが発生しました。",
     "input": "正しく入力されていない項目があります",
     "required": "を入力してください",
     "email": "正しい形式のメールアドレスを入力してください"
