+++ nuxt.config.js
@@ -16,6 +16,8 @@ export default {
     link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }],
   },
 
+  ssr: true,
+
   // Global CSS: https://go.nuxtjs.dev/config-css
   css: [{ src: '@/assets/scss/style.scss' }],
